+++ lib/entities/world_map.dart
@@ -1,0 +1,20 @@
+export '../src/entities/world_map/world_map.dart' show WorldMap;
+
+export '../src/generators/world_map/world_map_generator.dart'
+    show WorldMapGenerator;
+
+export '../src/subtypes/world_map/world_map_settings.dart'
+    show WorldMapSettings;
+export '../src/subtypes/world_map/world_map_settings_manager.dart'
+    show WorldMapSettingsManager;
+
+export '../src/subtypes/world_map/flatlands/flatlands_world_map_settings.dart'
+    show FlatlandsWorldMapSettings;
+export '../src/subtypes/world_map/islands/islands_world_map_settings.dart'
+    show IslandsWorldMapSettings;
+export '../src/subtypes/world_map/mountainous/mountainous_world_map_settings.dart'
+    show MountainousWorldMapSettings;
+export '../src/subtypes/world_map/normal/normal_world_map_settings.dart'
+    show NormalWorldMapSettings;
+export '../src/subtypes/world_map/rocky/rocky_world_map_settings.dart'
+    show RockyWorldMapSettings;