+++ lib/src/subtypes/world_map/world_map_settings.dart
@@ -1,0 +1,10 @@
+import '../../enums/terrain.dart';
+
+abstract class WorldMapSettings {
+  /// Returns the name of the setting
+  String getSettingName();
+
+  /// Returns the correct terrain type based on the height.
+  /// Height is between -1 and 1
+  Terrain getTerrainFromHeight(double height);
+}